Chat room can delete only the owner

+++ chat/src/main/java/greencity/service/impl/ChatRoomServiceImpl.java
@@ -185,12 +185,16 @@ public class ChatRoomServiceImpl implements ChatRoomService {
     }
 
     @Override
-    public void deleteChatRoom(ChatRoomDto chatRoomDto) {
-        chatRoomRepo.deleteById(chatRoomDto.getId());
-        Map<String, Object> headers = new HashMap<>();
-        headers.put(HEADER_DELETE_ROOM, new Object());
-        for (ParticipantDto participant : chatRoomDto.getParticipants()) {
-            messagingTemplate.convertAndSend(ROOM_LINK + participant.getId(), chatRoomDto, headers);
+    public void deleteChatRoom(Long ownerId, ChatRoomDto chatRoomDto) {
+        if (chatRoomDto.getOwnerId() == ownerId) {
+            chatRoomRepo.deleteById(chatRoomDto.getId());
+            Map<String, Object> headers = new HashMap<>();
+            headers.put(HEADER_DELETE_ROOM, new Object());
+            for (ParticipantDto participant : chatRoomDto.getParticipants()) {
+                messagingTemplate.convertAndSend(ROOM_LINK + participant.getId(), chatRoomDto, headers);
+            }
+        } else {
+            throw new UnsupportedOperationException(ErrorMessage.USER_NOT_THE_OWNER);
         }
     }
